Added MP camera support for unnatural corpse event
This won't be as smooth as it would be in SP due to camera data being only sent periodically, but it should end up with a closer experience to SP.

It should completely ignore the camera if those patches failed or are disabled.

This patch will remove any MP transpilers on the related methods, to remove MP patches from those methods. This is done to prevent MP from replacing what Rocketman MP does with this.